requirements.txt
Flask==2.3.2
Flask-CORS==3.0.10
PyPDF2==3.0.1
Pillow==9.5.0